Create separate test with append OOM
And add more tracers to base append test.
