+++ tests/CMakeLists.txt
@@ -104,7 +104,10 @@ add_test_generic(NAME stream_from_map TRACERS none memcheck pmemcheck drd helgri
 
 if(TESTS_RAPIDCHECK)
 	build_test_rc(NAME append SRC_FILES unittest/append.cpp)
-	add_test_generic(NAME append TRACERS none)
+	add_test_generic(NAME append TRACERS none memcheck pmemcheck)
+
+	build_test_rc(NAME append_oom SRC_FILES unittest/append_oom.cpp)
+	add_test_generic(NAME append_oom TRACERS none)
 
 	build_test_rc(NAME concurrent_iterate SRC_FILES unittest/concurrent_iterate.cpp)
 	add_test_generic(NAME concurrent_iterate TRACERS none)
