+++ tests/unittest/append.cpp
@@ -79,35 +79,6 @@ int main(int argc, char *argv[])
 				stream.sut.helpers.verify(region, {entry}, {});
 			});
 
-		ret += rc::check(
-			"verify append will work until OOM", [&](pmemstream_with_single_empty_region &&stream) {
-				auto region = stream.sut.helpers.get_first_region();
-
-				size_t elems = 10;
-				const size_t e_size = TEST_DEFAULT_REGION_SIZE / elems - TEST_DEFAULT_BLOCK_SIZE;
-				std::string e = *rc::gen::container<std::string>(e_size, rc::gen::character<char>());
-
-				struct pmemstream_entry prev_ne = {0};
-				while (elems-- > 0) {
-					auto [ret, new_entry] = stream.sut.append(region, e);
-					UT_ASSERTeq(ret, 0);
-					UT_ASSERT(new_entry.offset > prev_ne.offset);
-					prev_ne = new_entry;
-				}
-				/* next append should not fit */
-				auto [ret, new_entry] = stream.sut.append(region, e);
-				UT_ASSERTeq(new_entry.offset, prev_ne.offset);
-				/* XXX: should be updated with the real error code, when available */
-				UT_ASSERTeq(ret, -1);
-				e.resize(4);
-				/* ... but smaller entry should fit just in */
-				std::tie(ret, new_entry) = stream.sut.append(region, e);
-				UT_ASSERT(new_entry.offset > prev_ne.offset);
-				UT_ASSERTeq(ret, 0);
-
-				UT_ASSERTeq(stream.sut.region_free(region), 0);
-			});
-
 		ret += rc::check("verify if sequence of append and reopen commands leads to consitent state", [] {
 			pmemstream_model model;
 			pmemstream_with_single_empty_region s;
